통신 추가 및 마이페이지 갱신 로직 추가

+++ FLUFFY/Extension/UserDefaults+Ext.swift
@@ -43,9 +43,9 @@ extension UserDefaults {
         }
     }
     
-    var userScore: Int {
+    var userScore: Double {
         get {
-            return UserDefaults.standard.integer(forKey: Const.kIsUserScore)
+            return UserDefaults.standard.double(forKey: Const.kIsUserScore)
         }
         set {
             UserDefaults.standard.set(newValue, forKey: Const.kIsUserScore)
